Add VSCode conda terminal profiles
Since this project uses conda environments, it is required that the
conda command is available in the terminal. In Windows, this is done via
the anaconda prompt. This commit creates two terminal profiles in VSCode
to launch the appropriate terminal such that the conda command is
available. As the default the miniconda is assumed to be installed but
the profile for anaconda is also provided requiring a manual change in
the settings.json file

Refs: #1